TG now uses fistkind
because a heart player definitely knows how to use their self as a weapon.
(also because i almost always go the unarmed route in a video game, if it's viable).
and yeah, i know this means TG can have his fists stolen... and upgraded. but that's not a bug, it's a feature!